A proof-of-concept implementation of just-in-time ContextDefinition loading.
This may need refactoring, but the contract with consumers should not change:
SubContexts will be parsed even if added after the initial parsing.

Before this commit, subcontexts were only parsed if they were added to a context
during it's first initialization. Now, a parse is requested just prior to
looking up a step match. This means, most importantly:

Subcontexts can be added during a test sequence.

Notably, subcontexts can be added conditionally during hooks!

This is acheived by attaching the AnnotatedLoader to the DefinitionDispatcher,
which then requests a parse just prior to seeking a definition.
For redundency avoidance, the loader keeps track of previously loaded classes,
and does recursive (subcontext) loading itself.

I fully acknowledge that there is a better way to do this in the Symphony framework.
I simply didn't have the time to play nice in a new framework before I need this.
I think a big fail is that the ContextReader is bound to only the 'main' context
and is only doing it's reading early on. The basic idea of this commit should
make it's way into the core mechanisms.

An ideal alternative would be an observer pattern, where the Loaders observe
Contexts (anything they have previously loaded) for reload-type events; and
adding a subcontext is a perfect example of that.