Prepare a pass for typehinted arguments
this is done in order that the next prepare call
(prepareNumberedArguments) have the right indexes for its arguments.

E.g, if for 5 arguments, we had the 2 typehinted `[1 => Foo, 3 => Bar]`,
before it would try to replace those with `[0 => 'foo', 1 => 'bar', 2 =>
'baz']`, resulting in `[0 => 'foo', 1 => Foo, 2 => 'baz', 3 => Bar]`
which is wrong, as we would expect it to be `[0 => 'foo', 1 => Foo, 2 =>
'bar', 3 => Bar, 4 => 'baz']`

To achieve that, we need to indicate that the typehinted args are
defined... Thus they are indeed skipped when trying to match numbered
args.