Simplify installation instructions
At the time these instructions were written it might've been necessary to explain how to install composer. With composer becoming a de-facto standard package manager for PHP I think it's now better to assume it's installed, and refer to composer's installation docs. The installation method we have here is outdated (and insecure) anyway.