Fix fatal due to wrong symfony/event-dispatcher version detection
The version detection added in #1256 to support symfony 5 relies
only on checking for the presence of a class provided by the
(separate) symfony/event-dispatcher-contracts package.

However, symfony/console 4.2 also requires symfony/contracts
but can coexist with older versions of symfony/event-dispatcher.

Therefore for example if a project has a constraint on updating
symfony/event-dispatcher (e.g. because of using Guzzle 3,
which requires event-dispatcher ~2.1) composer will install:

* symfony/console @ 4.2.12 (latest that doesn't advertise a
  conflict with newer event-dispatcher)
* symfony/contracts @ 1.1.8 (latest that satisfies the console
  requirement)
* symfony/event-dispatcher @ 2.8.52 (latest that satisfies
  guzzle)

In this setup the presence of symfony/contracts incorrectly
signals to behat that we are running symfony 5 and causes a
fatal error when it loads the new implementation of
TestworkEventDispatcher.

The only change between the two versions is that the method
signature for dispatch has changed order and types.

Instead of looking at packages, it is more robust just to use
reflection to detect which version of the dispatch() method
is defined, and then load the extension class that matches
that signature.

Fixes #1277