Scenario Hook should allowed neated filters (fixes #1235)
Scenario hooks support filtering by tag. The tag can be filtered on
either the Feature tags, or the Scenario tags.

When checking these tags, the tags must be combined before checking to
allow support of negated tags.

The `TagFilter::isScenarioMatch()` function in Gherkin takes both the
FeatureNode, and ScenarioNode as arguments, and then combines the tags
on both of these.

This fix addresses an issue whereby the Feature does not contain a tag,
but the Scenario does, and the hook is filtered on the absence of that
Scenario tag.

For example, Given the following Feature:
```
Feature:

  @doNotMatch
  Scenario:
    I should not be called with the @~doNotMatch filter
```

And the following Hooks:
```
/**
 * @BeforeScenario @~doNotMatch
 */
public negatedBeforeScenarioHook($event) {
  throw new \Exception('I should not have been called');
}
```

The Feature contains no hooks, but the Scenario contains the
`@doNotMatch` hook.

When processed before applying this patch the call to
`TagFilter::isFeatureMatch($feature)` will not not match (i.e. it _will_ match),
the `isMatchTagFilter` function returns early with a true value, and
therefore the hook incorrectly applies.

Following this change the, `TagFilter::isScenarioMatch($scenario)`
function is called and checks the values of all tags on _both_ the
Feature, _and_ the Scenario. The tag does _not_ match, the
`isMatchTagFilter()` function return false,a dn the tag _does not_ apply
which is the correct behaviour.