Add support for specification of a prefered profile
In many situations it is desirable to have a specific profile that is
used when none is specified, however because of the way in which all
profiles inherit from the `default` profile, this is not currently
possible.

For example, if you have different MinkExtension configurations for a
range of different browsers, then each browser may only be specified in
its own profile. This works well, but often you will want to specify a
default browser configuration.

To achieve this you may think that you can set the Extension
configuration for your preferred browser into the `default` profile, but
this means that all other profiles will include this configuration, and
may not work as expected.

This change introduces the concept of a preferred profile, which is the
profile which will be used unless the `--profile` CLI argument is
provided to override it.

This means that if you have profiles for chrome, headlesschrome,
firefox, headlessfirefox, edge, and headlessedge, you can choose to
specify a preferred profile of `headlesschrome`, which will be used
unless you explicitly set the `--profile` argument when running behat.