Fix CI fails because symfony/flex is blocked
# Problem:
When running the workflow, it fails at the "Install symfony/flex" step with the following error:

Error: symfony/flex contains a Composer plugin which is blocked by your allow-plugins config. You may add it to the list if you consider it safe.

## Cause:
A newer version of Composer has introduced a strict check whereby all plugins have to be explicitly allowed for security reasons.

## Fix:
Explicitly allow the global plugin before installing it.