GH Actions: start running the tests against PHP 8.3
PHP 8.3 is nearing the RC phase. As this tool is used in the build processes of plenty of other PHP package, I believe it would be beneficial to start supporting PHP 8.3 officially sooner rather than later.

This commit adds PHP 8.3 to the build matrix to start running the tests against PHP 8.3.

As the build currently passes against PHP 8.3, I've not added a `continue-on-error` for PHP 8.3.