Update box-config
This removes some defaults that are no longer necessary as well as the
files config that stopped Box from applying it'S own logic of figuring
out what is necessary.

This should now produce a working PHAR file that is similar to the one
based on box2 but now with the latest box-releases so that bugfixes are
much easier to be applied.

This still is - similar to the previous version - not a scoped build so
that there might be circumstances where the executed code uses code from
the PHAR file instead of from their own code. This though is a separate
issue.