Move PHAR build into separate build-step
This commit moves building the PHAR file out of the test-suite and moves
it into it's own step that requires all biulds to pass to be executed.
That makes it for one thing easier to make sure that all tests in all
versions of PHP work before creating the PHAR and it also allows to
easier optimize the build of the PHAR without interfering with tests.

The build-steps from the test-part were taken over into the build-phar
part and appropriately extended.

Also this makes sure that the PHAr is build based on the PHP7.2 version
so taht it can be ran in all supported PHP-Versions.