Remove deprecations in PHP8.4
PHP8.4 has deprecated implicit nullable parameters. Some of those are
used within Behat and now render a lot of deprecation messages.

This PR replaces the implicit nullables with an explicit nullable
wherever they are encountered. This should not have any influence on the
code-execution but merely updates the function signature to make more
explicit what is expected.