PHP 8.4 | Fix passing E_USER_ERROR to trigger_error()
PHP 8.4 deprecates passing `E_USER_ERROR` to `trigger_error()`, with the recommendation being to replace these type of calls with either an Exception or an `exit` statement.

This commit fixes the one instance found in this codebase.

Based on my reading of this test, the test is not so much about the error _level_, but about the error _handling_.
Given that, changing the error level from `E_USER_ERROR` to `E_USER_WARNING` preserves the value of the test while avoiding the PHP 8.4 deprecation notice.

Ref: https://wiki.php.net/rfc/deprecations_php_8_4#deprecate_passing_e_user_error_to_trigger_error