Remove running behat against behat
That causes some odd sideeffects when trying to run the tests inside the
phar.

This needs more investigation and a separate PR. The main reason for
testing the PR was testing that no immediate problems arise in the PHAR
file. That should already be tested when running the `--version` command