change: deprecate ApplicationFactory::VERSION, auto-detect behat version
The ApplicationFactory::VERSION constant has not been bumped for the
last couple of releases, and updating it manually every time is an
avoidable maintenance burden.

Deprecate the constant - it will not be updated in the future, and will
be removed in next major.

Instead, use composer's runtime API to detect the version at runtime
(only really used for inclusion in tool output). This follows the
pattern used by e.g. doctrine
https://github.com/doctrine/orm/blob/44fa8bbde87238ad9b574e87325c7ff89c75837e/src/Tools/Console/ConsoleRunner.php#L45-L46

Composer's version is generated to a file at the time of `composer install`
and therefore is baked into the `.phar` so that also works as expected.

Note that in some cases composer versions are non-numeric (e.g. a dev branch
may report a version of `dev-c629b4a773cf2dc11e0930bd8930adf22eae5031`).