fix: Config converter should use `->addContext` for contexts with args
Previously, it was assuming that all `contexts` values would be simple
string class (or service) names that could safely be passed to
`->withContexts(string ...$contexts)`.

However, it is also possible to define contexts with constructor
arguments. These need to be passed individually to `->addContext()`.

YAML files can contain contexts with and without arguments in an
interspersed list. When converting that to PHP, we need to ensure the
order that the contexts are loaded does not change (to guarantee the
same behaviour of any hooks etc that the user is registering).

Therefore, if *any* context takes an argument, *all* contexts will be
loaded with`->addContext` (rather than attempting to split between
`->withContexts` and `->addContext`).