feature: Import & use ::class references when converting YAML config
Extensions and Contexts will often (but not always) be configured with
a fully qualified PHP class name. Rather than converting these to PHP
strings, we can convert them as e.g. `FeatureContext::class` references
with a suitable `use` statement.

This will get the resultant file closer to what we'd encourage writing
for new projects.

For both Extensions and Contexts, Behat supports a variety of references
including service locators and file paths. Therefore, we only convert an
argument if it passes a `class_exists()` to ensure that it is definitely
a class name.