test: Prove that env vars in config are converted correctly
It would theoretically be possible to use `getenv()` etc directly in the
PHP config. However, the symfony placeholder implementation also works
and would require a bit of thought to convert to native PHP code. This
is best left to end-users to consider the pros/cons & handle manually.

But given environment variables may contain secrets etc, I think it's
worth having the feature coverage prove that these are converted as
the placeholder (rather than the value).