feature: Skip default method arguments when converting config to PHP
This makes the converted configuration cleaner particularly for
boolean methods like `withStopOnFailure()`.

This commit also makes the config conversion process stricter - it will
now fail if a config object attempts to generate a method that does not
exist, or provides arguments that do not match the method signature.
This will help us to keep the `->toPhpExpr` in sync with any changes to
the config objects over time.