feature: Convert error_reporting values to constants in PHP config
If the user has defined error_reporting integers in their YAML
config files, we can attempt to convert these to appropriate PHP
constant expressions. This will make the configuration clearer
(and help with the fact that the user may have a comment in the
YAML explaining the value, which will be lost in conversion).