fix: ExceptionPresenter was causing a TypeError with null $basePath
As reported in #1630 and #1629, we missed that the ExceptionPresenter
can be validly created with a null $basePath but it is not valid to
pass this to the new ConfigurablePathPrinter. Although Behat never
does this, third-parties do.

The old behaviour in that case was that the ExceptionPresenter would
just print absolute paths.

We can achieve the same thing by setting the correct value for
printAbsolutePaths when we create the ConfigurablePathPrinter. We
can then just cast the $basePath to an empty string as this will
be ignored if we're printing absolute paths.