refactor: Consolidate responsibility for generating step method names
Previously, most of this was done within duplicated methods in the
`PatternPolicy` implementations and passed into the `Pattern` as
`canonicalText`.

There are lots of potential interpretations of the `canonical text` of a
step but in practice we were only using this as the basis of the
suggested method name when generating snippets. This suggestion was
processed almost entirely in the `PatternPolicy` class with one final
step in the `ContextSnippetGenerator` to lowercase the first character.

It makes more sense to consolidate all that logic together, and to make
explicit that this property on the `Pattern` is the suggested method
name.