change: Ensure suggested method names are valid for PHP
Just remove any characters that are not within the allowed set
for method names (including digits at the start of the name),
and fall back to the generic `stepDefinition1` if the text
is completely invalid.

Since we may now have utf8 characters in the suggestion, we
also need to make the case conversions (to lowerCamelCase)
multibyte-safe. Fortunately, Behat already depends on
ext-mbstring.