change: Don't clean up step name suggestions from 3rd-party PatternPolicy
The previous implementation introduced a new-but-deprecated method to
attempt to handle the (unlikely) case that a user has a custom
PatternPolicy suggesting step method names.

In practice, the only thing that is actually *required* to produce
valid code is that the name is not empty. We can easily handle that
without extracting a complete internal method.

Worst case, if a user's custom code is not lowercasing the first
character of their method name, we may generate a method that does
not follow their coding style. This feels like a reasonable trade-off
for the improvement to code clarity / structure here.