test: Switch i18n syntax help feature to greek to fix gherkin compat
Behat/Gherkin@4.13.0 introduced a number of new translations. These are
BC for end-users (when parsing the features, any of the past-or-present
translations will be recognised).

However, our `--story-syntax` command outputs all known translations
when displaying the example feature. This means the expected output in
our tests has to exactly match the Behat/Gherkin translations across all
supported versions. This caused CI to break due to extra Russian
translations between 4.12 and 4.13.

I have switched the test to Greek because it appears to be stable (has
not changed in cucumber/gherkin in the last 8 years), still contains
UTF-8 to prove character sets work as expected, and has a variety of
keywords with/without synonyms.