chore: Relax constraint on composer/xdebug-handler
These will allow users to upgrade to newer Behat versions even
if they have a dependency that is holding back their
composer/xdebug-handler version.

* 1.4.0 adds support for PHP 8 and is therefore the minimum we can allow
* 2.x has API changes, but none that affect the methods we are calling
* 3.x only drops support for old PHP and adds strict types

Therefore, it should be safe to accept all of these versions.

Note, however, that we have no easy way to test this in CI due conflicts
with our own dev dependencies, and because the main Behat process
disables xdebug and persists this config to children - so cannot easily
run child Behat processes with xdebug enabled.