chore: Support 4.19.x versions of nikic/php-parser
Requiring ^5.2 was causing dependency conflicts that were preventing
end-users from upgrading to newer Behat versions.

We only actually *require* nikic/php-parser for running the
convert-config command to generate PHP config from the old YAML files.
I therefore explored making it a suggested dependency that was only used
to run that command.

The problem is that it is also required indirectly by many of our dev
dependencies (including PHPUnit, which we use to run our assertions).
This means we can't configure a CI run where it is **not** installed,
so would not have confidence that Behat does indeed work as expected
without it.

The best compromise is to relax the constraint to ^4.19.2 (this version
introduced enum support, which we need). That should resolve the
conflict for the majority of users.