test: Define behaviour when step cannot be represented as turnip
A step may contain literal punctuation that cannot be safely represented
in a turnip pattern - since there is no official strategy for escaping
the characters that turnip uses for placeholders etc.

Until now, we have generated snippets that do not actually match the
step text. This means these snippets are generated repeatedly and
ultimately then cause an exception on the next run due to the presence
of multiple snippets with the same pattern.