refactor: Move responsibility for matching pattern to step text
Previously, the RepositorySearchEngine asked the PatternTransformer
to convert each individual pattern to regex before itself calling
preg_match to attempt to match the pattern.

Instead, move the responsibility for the matching the pattern
(converting and testing it) to the PatternTransformer. This will make it
easier to reliably reuse the same logic to test that a step *actually*
matches the generated pattern when generating snippets.