test: Fix assertions on removed file paths when converting config
These assertions were checking for deletion of the wrong file paths for
the scenario. Although the step is worded "should have been removed"
it is in practice only checking that the file does not exist - and
therefore was passing because the files had never been created.

Updated to the correct names of the files.