test: Run fixture-based scenarios in isolated directories
Previously, we were running the scenario in the fixture directory itself
(or explicitly copying individual files to a temporary directory). This
risks Behat modifying fixture files during the scenario (potentially
unexpectedly) and therefore affecting subsequent scenarios.

Instead, run every scenario in an isolated temporary working directory,
and simply initialise this from the fixture at the beginning. This
guarantees that every scenario runs in an isolated environment.

It also simplifies our step definitions, as we no longer need to provide
for copying individual files to a temp dir or having separate steps to
assert the content / presence of files in the temp dir.

Placeholders to some paths in features have changed because the paths
are now relative to the %%WORKING_DIR%% placeholder (and are not nested
under the fixture directory path) rather than to the project root.