test: Use realpath to workingDir - it may not be the same
On MacOS and Windows runners, the path returned by sys_get_temp_dir
(and stored in the workingDir variable) is not a real path. Therefore,
it does not match when passed as an option (e.g. for --remove-prefix)
into Behat. We need to convert to a realpath so that it matches the
actual path to the files on disk.