fix: Ensure that suggested step definitions actually match the step text (#1656)
Previously, when generating snippets for undefined step definitions,
there was no guarantee that the pattern would actually match the step
text. In some cases (in particular with turnip) this proved to be an
issue when feature files contained complex or unexpected syntax:

* The first run would apparently provide a valid snippet.
* The second run would still not match this snippet, so would propose a
  new step with the same pattern.
* The third run would fail due to the presence of duplicate patterns in
  the steps from the first and second run.

Instead, check that the generated snippet actually matches the step text
at the time of generating it. If not, show the user a warning with
information about either generating these as regex or manually defining
a step to match their features.

Fixes #1653