feat: throw if a step provides an unexpected PyString or Table argument (#1614)
Previously, if a step definition in a feature file had a multiline argument
(e.g. PyString or Table) that was not used by the step definition, it would
be silently ignored.

This could cause false passes if for example a feature was using incorrect
wording or a line had been accidentally deleted / commented.

Behat will now throw an InvalidArgumentException to highlight that the
definition of the step does not match the usage.

Fixes #1609 according to [@acoulton
comment](https://github.com/Behat/Behat/issues/1609#issuecomment-2769851124)