feat: Display line number of failing test in pretty formatter (#1615)
Previously, when printing information about failures, the pretty
formatter only displayed the line number of the start of the 
failing scenario. In a long scenario, this might be quite 
separated from the line number of the actual step that failed.

It might also be *after* the line of the failing step, when 
working with Scenario Outlines (where the scenario is counted
as starting on the line containing the relevant example table
row).

Now, Behat will also print the line number of the failed step.
This makes the behaviour equivalent to the progress 
formatter, which already provided this information.

Fixes #1612