fix: Provide more context when failing with unexpected multiline args
Because of the stage where it is thrown, the exception is treated as
an overall fatal error. This meant that Behat does not output any detail
on the feature file / step it was parsing when it encountered the
unexpected argument. As a result, users had no information on the cause
of the error and would have to check their features / steps manually.

Instead, include the name of the function and (where possible) the
feature and step within the exception message. This will still mean that
users have to fix these one at a time (Behat will exit on the first such
exception) but they should at least now be able to locate the issue.