Implement JSON formatter (#1674)
Lately JSON has become the lingua franca of tools, which use this format
to exchange information everywhere. This PR introduces a JSON formatter
which is able to save Behat test results as a JSON file. One possible
application for this format is to be able to create an MCP server for
Behat, using the generated JSON to return information to the client

This formatter is heavily inspired in the existing JUnit formatter and
the exported output is similar but, since it is not constrained by the
existing JUnit specification, it follows more closely the structure of
our project. The generated JSON basically will contain a list of Suite
nodes, which will contain Feature nodes, which will contain Scenario
nodes. Each node will have properties indicating some info like the
status, the number of tests for each status, file and line and execution
time. Also, the JUnit formatter produced one file per suite, this
formatter produces just a single JSON file for all suites so the
`output_path` needs to be a file name, not a folder name

The PR also defines a JSON schema which can be used by the tools that
will use this information