Use configurable path printer in JSON and JUnit formatters (#1677)
The JSON and JUnit formatters were using a very simple management of the
printed file paths. This PR replaces this with the
ConfigurablePathPrinter that the other formatter use. This brings a
couple of advantages:
- It uses the `%paths.base%` to calculate relative paths, this is more
correct than using cwd as those formatters were doing
- It allows us to use the options offered by that printer

The editorUrl option of that printer is not appropriate for these
formatters, which cannot use the link added to the paths, so I added a
flag to mark if this option needs to be applied or not

Adds tests that check the correct working of the printer options in
these formatters