Fix setting error_reporting (#1675)
When the error reporting level was set (either by using
`withErrorReporting()` or by using the default value of E_ALL), we were
calling `set_error_handler()` with this level but were not calling
`error_reporting()`. This meant that event though we had set a handler
to be called on certain error levels, this handler might not be called
if the error_reporting level set in php.ini was more restrictive. I saw
this in a system where error_reporting was set to `E_ALL & ~E_DEPRECATED
& ~E_STRICT` and the test `With default error reporting reports all PHP
errors and deprecations` failed because the deprecation was not raised

This PR fixes this problem by actually setting the error_reporting
level, not just using it in the set_error_handler call

Fixes https://github.com/Behat/Behat/issues/1663