fix: Include consistent index & placeholders for Example names in junit
Previously, when a feature file contained scenarios with `Example` nodes
the junit report would append an index to each testcase name to
differentiate them (e.g. `<testcase name="Some scenario outline #1"`,
`<testcase name="Some scenario outline #2`, etc).

These indices were dynamically generated based on the examples that
actually ran. As a result, on a `--rerun` the names would start from
`#1` in sequence and be inconsistent with the numbering of the first
execution. This makes it difficult to compare reports between each
test run, for example to link a passing rerun to the previous
failure.

With this change, examples are numbered consistently based on the
order they appear in the feature file.

Additionally, we also now support placeholders in the Scenario
Outline title. This allows users to include some or all of the
values from the Examples: table in each `<testcase` name for
additional clarity. This matches the behaviour of the official
cucumber runners.

This PR builds on work done by @uuf6429 and in particular the
implementation of the numbering & placeholder replacement that
he contributed to Behat/Gherkin.