Add "timer" option to JSON and JUnit formatters (#1681)
The information that shows the time spent to run each test in the JSON
and JUnit formatters is interesting if you are trying to optimise the
performance of your tests but in many other cases you may only be
interested in whether the tests pass or not and this timer information
will be of no interest to you. This PR adds a "timer" option to these
formatters (similar to the one already available in the pretty and
progress formatters) that can be used to disable the output of this
data. It defaults to true to preserve existing functionality