Check JSON file before running all tests (#1684)
With the current implementation, the checks for the JSON file (check
that the output path had been passed, that it was not a directory, that
we could create the file) would not happen until after all the tests had
been run, which could be very inconvenient, run all tests only to find
that you cannot write the file.

In this PR we change this to follow what the JUnit formatter does, which
is to try to create the file at the beginning of the test run.