fix: FixedContextIdentifier should correctly handle null contextClass
`FixedContextIdentifier` is always created and registered by the
`ContextSnippetsController` with an injected `contextClass` based on the
 `--snippets-for` CLI option - see
https://github.com/Behat/Behat/blob/0d4162bbf8401bfba8085a25f427eabe6d00cead/src/Behat/Behat/Context/Cli/ContextSnippetsController.php#L65.

If this option has not been specified on the CLI then the `contextClass`
property will be `null`.

This was causing us to call `ContextEnvironment::hasContextClass` with
a `null` despite that method being typed as taking only
`class-string<Context>` (the method currently only has a phpdoc
typehint). This means `InitializedContextEnvironment::hasContextClass()`
then attempts to use `null` as an array key.

Up to PHP 8.4 this worked as expected, but from 8.5.0 it triggers a
deprecation.

Therefore, we now explicitly check if the value is null in
`FixedContextIdentifier` to ensure that we only call the
`->hasContextClass()` with a valid string value.