Update the pretty printer to handle future gherkin tag behavior (#1690)
The legacy mode of the gherkin parser was trimming the `@` sign from
tags when building the AST, so the pretty printer was prepending it
back.

As the upstream cucumber gherkin AST preserve the `@` sign in tags,
future versions of the behat/gherkin parser will preserve it as well.

Refs https://github.com/Behat/Gherkin/issues/336

The pretty printer can handle that case by detecting the format of the
tag, without any need of knowing the compatibility mode used by the
parser.
Note that the filtering by tags is not updated in this PR because the
implementation of the filter is done in behat/gherkin, not in this repo.