feat: Add constructor interface for SimpleArgumentTransformation
When the TransformationCalleeFactory creates instances of classes that
implement SimpleArgumentTransformation, it assumes that they all take
the same constructor parameters. This was not previously explicit in the
interface, which caused confusion and produced phpstan errors if a
transformation class accepted the arguments but did not use them.

Added the constructor to the interface to ensure that all
implementations of this class can be created as expected.