test: Prove that Scenarios support multiple Examples tables
Feature writers should be able to structure examples into multiple
tables, so that they can be clearly grouped into different kinds
of values.

Historically, this was not supported - see #1082 - and was
believed to be a BC break in Behat.

However, it was later implemented via a workaround in
Behat/Gherkin#161 which was released in 2021.

I have added test coverage to prove that features like this are
executed as expected. As can be seen, the current pretty printer
output is not perfect (we lose the detail of the separate tables
in the original file) however it looks like changing this would
be a BC break due to the way the pretty printer is implemented.

Fixes #1082 - I will open a separate issue to improve the pretty
printer output with example tables (which may also relate to the
upcoming Gherkin improvements to support descriptions on these
nodes).