[4.x] feat: Add Symfony 8 support (#1687)
This PR updates version constraints on `symfony/*` packages to allow
Symfony 8, which will be released this month.

This required adding strict return types to a number of methods
for compatibility with the changes to the underlying interfaces in
symfony 8. End-users will similarly have to add return types in their
own projects as part of the upgrade to Behat 4.0.