Fix static analysis in PHP 8.5 (#1719)
The latest versions of the Symfony Config component improve the docs for
some of the functions on the config objects and now use Template tags so
that PHPStan is able to correctly guess the type returned. So some
asserts are not needed any more. But they are still needed in older
versions of Symfony which don't have these improved docs. The only
solution I was able to find was to change these asserts into @var tags,
this seems to work for both versions

Had to disable a PHP-CS-Fixer rule which was incorrectly trying to
change these @var tag docblocks into normal comments